[feat] HeroSection 컴포넌트 렌더링 최적화, Slide 컴포넌트 구현
