+++ src/pages/Home/HeroSection.tsx
@@ -1,30 +1,18 @@
-import { useEffect, useState } from 'react';
+import { useNavigate } from 'react-router-dom';
 
 import Button from '@components/Button';
 import Section from '@layouts/Section';
 import { mainBackground } from 'src/data/main';
 
-import SlideDot from '@assets/img/slide-dot.svg?react';
 import ScrollArrow from '@assets/img/scroll-top-arrrow.svg?react';
-import { useNavigate } from 'react-router-dom';
+import SlideDot from '@assets/img/slide-dot.svg?react';
+import Slide from './Slide';
 
-const DURATION = 3000;
 const ANIMATION_DURATION = 'duration-[2s]';
 
 const HeroSection = () => {
   const navigate = useNavigate();
-  const [index, setIndex] = useState<number>(0);
-
-  const intervalBackground = (clidkedIndex: number) => setIndex(clidkedIndex);
   const moveToCommunicationPage = () => navigate('/qna/communication');
-
-  useEffect(() => {
-    const intervalId = setInterval(
-      () => (index < 2 ? setIndex(index + 1) : setIndex(0)),
-      DURATION,
-    );
-    return () => clearInterval(intervalId);
-  });
 
   return (
     <Section
@@ -45,62 +33,67 @@ const HeroSection = () => {
         </span>
         <ScrollArrow className="lg:h-auto sm:h-[100px] rotate-180" />
       </div>
-      <div className="relative lg:w-auto sm:min-w-[440px]">
-        {mainBackground.map(({ imgName, imgSrc }, bgIndex) => (
-          <img
-            key={imgName}
-            className={`
-              ${index === bgIndex ? 'opacity-100' : 'opacity-0'}
-              xl:w-[1338px] lg:w-[785px] sm:w-[440px]
-              absolute transition ${ANIMATION_DURATION}
-              xl:-top-[23rem] xl:left-[15.625rem]
-              lg:-top-[10rem] lg:-right-[5.375rem]
-              sm:top-[0.875rem] sm:-right-[1.5rem]
-              -z-50
-            `}
-            srcSet={`${imgSrc[2]} 1024w, ${imgSrc[1]} 1920w, ${imgSrc[0]}`}
-            alt={imgName}
-          />
-        ))}
-        <article className="xl:pt-[23.25rem] lg:pt-[12rem]">
-          <div className="lg:flex sm:hidden items-center gap-3.5 ">
-            {mainBackground.map(({ imgName }, bgIndex) =>
-              index === bgIndex ? (
-                <SlideDot key={imgName} />
-              ) : (
-                <button
+      <div className="relative lg:w-auto sm:min-w-[440px] xl:pt-[23.25rem] lg:pt-[12rem]">
+        <Slide duration={3000}>
+          {({ index, changeIndex }) => (
+            <>
+              {mainBackground.map(({ imgName, imgSrc }, bgIndex) => (
+                <img
                   key={imgName}
-                  aria-label={`${imgName}-button`}
-                  className="inline-block rounded-full w-2.5 h-2.5 bg-gray-7"
-                  onClick={() => intervalBackground(bgIndex)}
+                  className={`
+                    ${index === bgIndex ? 'opacity-100' : 'opacity-0'}
+                    xl:w-[1338px] lg:w-[785px] sm:w-[440px]
+                    absolute transition ${ANIMATION_DURATION}
+                    xl:-top-[23rem] xl:left-[15.625rem]
+                    lg:-top-[10rem] lg:-right-[5.375rem]
+                    sm:top-[0.875rem] sm:-right-[1.5rem]
+                    -z-50
+                  `}
+                  srcSet={`${imgSrc[2]} 1024w, ${imgSrc[1]} 1920w, ${imgSrc[0]}`}
+                  alt={imgName}
                 />
-              ),
-            )}
-          </div>
+              ))}
 
-          <h1
-            className="xl:pt-[3.75rem] lg:pt-20 sm:pt-[21.25rem]
+              <div className="lg:flex sm:hidden items-center gap-3.5 ">
+                {mainBackground.map(({ imgName }, bgIndex) =>
+                  index === bgIndex ? (
+                    <SlideDot key={imgName} />
+                  ) : (
+                    <button
+                      key={imgName}
+                      aria-label={`${imgName}-button`}
+                      className="inline-block rounded-full w-2.5 h-2.5 bg-gray-7"
+                      onClick={() => changeIndex(bgIndex)}
+                    />
+                  ),
+                )}
+              </div>
+            </>
+          )}
+        </Slide>
+
+        <h1
+          className="xl:pt-[3.75rem] lg:pt-20 sm:pt-[21.25rem]
               xl:text-[5rem] lg:text-[3.75rem] sm:text-[2.5rem]
               xl:leading-[6.25rem] lg:leading-[4.75rem] sm:leading-[3.125rem]
             text-white font-bold mix-blend-difference"
-          >
-            어린이집 · <br className="lg:hidden" /> 유치원 선생님의 <br />
-            소통 공간
-          </h1>
+        >
+          어린이집 · <br className="lg:hidden" /> 유치원 선생님의 <br />
+          소통 공간
+        </h1>
 
-          <p
-            className="
+        <p
+          className="
               xl:pt-[1.875rem] lg:pt-2.5 sm:pt-4
               lg:pb-[3.125rem] sm:pb-10
               xl:text-[1.875rem] lg:text-2xl sm:text-[1.125rem]
               xl:leading-[2.875rem] lg:leading-10 sm:leading-7
             text-white mix-blend-difference"
-          >
-            <b>유어티처</b>는 <b>어린이집 · 유치원 선생님</b>을 위한 <br />
-            커뮤니티 플랫폼입니다.
-          </p>
-          <Button onClick={moveToCommunicationPage}>소통방으로 이동</Button>
-        </article>
+        >
+          <b>유어티처</b>는 <b>어린이집 · 유치원 선생님</b>을 위한 <br />
+          커뮤니티 플랫폼입니다.
+        </p>
+        <Button onClick={moveToCommunicationPage}>소통방으로 이동</Button>
       </div>
     </Section>
   );
